Move scaleIconOnHover to Icon
