+++ packages/react/src/call-to-action/CallToAction.tsx
@@ -1,8 +1,9 @@
 import { type ElementType, type PropsWithChildren } from 'react';
 import { type LinkProps } from '../core';
 import { Flex } from '../flex';
-import { BaseButton, scaleIconOnHover, type BaseButtonProps } from '../button';
+import { BaseButton, type BaseButtonProps } from '../button';
 import { ChevronRightIcon } from '../icon';
+import { scaleIconOnHover } from '../icon/Icon';
 import { TextLink } from '../text-link';
 
 export type CallToActionLinkProps = LinkProps;
